Added type definitions for signal types that are used multiple times (#79)
This is a little request for comments. 
I felt it would make sense to introduce some type definitions to reduce
the amount of duplication in the specification. Especially for more
complex definitions like the registerflags, this allows good reuse
without having to define it over and over. Seemed a better solution to
me over the previously defined derived parameter.
This adds hyperlinks where the types are used that take you to the type
definition.


![image](https://github.com/openhwgroup/core-v-xif/assets/101172126/06531072-cc27-42b0-9ef1-952709baa576)