Upgrading license from SHL-0.51 to SHL-2.1 (#135)
This PR is to change the license to the specification. This will
harmonize the license of the documentation with other components of this
project, e.g. around the document generation, and eventually the .sv
example. All reviewers, please give the green light for this change. I
would like to get this completed before sending the spec out for review.